requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap3==12.0.3
django-ckeditor==5.9.0
django-discover-runner==1.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-role-permissions==2.2.1
django-slack==5.14.3
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
lxml==4.5.0
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
python-docx==0.8.10
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
